% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{period_input}
\alias{period_input}
\alias{period_server}
\title{Period Selector}
\usage{
period_input(id, label = NULL, width = "480px")

period_server(
  id,
  min = "1900-01-01",
  max = "2100-01-01",
  start = min,
  end = max
)
}
\arguments{
\item{id}{The id of the module}

\item{label}{The label of the date range input}

\item{width}{The width of the date range input and shortcut button group}

\item{min}{The minimum allowed date. Either a Date object, or a string in
\code{yyyy-mm-dd} format.}

\item{max}{The maximum allowed date. Either a Date object, or a string in
\code{yyyy-mm-dd} format.}

\item{start}{The initial start date. Either a Date object, or a string in
\code{yyyy-mm-dd} format. If NULL (the default), will use the current
date in the client's time zone.}

\item{end}{The initial end date. Either a Date object, or a string in
\code{yyyy-mm-dd} format. If NULL (the default), will use the current
date in the client's time zone.}
}
\value{
The \code{input} function creates a date range input with shortcut buttons

The \code{server} function returns a reactive date range
}
\description{
Create a date range input with shortcut buttons for MTD, QTD, YTD, 1M, 3M, 6M, 1Y, 3Y, 5Y, 10Y,
and ALL. The dates can be set by selecting from the calendar, typing in the date, or clicking
shortcut buttons.
}
