% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procatapp-package.R
\docType{package}
\name{procatapp-package}
\alias{procatapp}
\alias{procatapp-package}
\title{procatapp package}
\description{
ProCat enables interactive process attribution for 4F funds. If you can't measure it, you can't
manage it!
}
\author{
Grant Irvine-Smith \email{grant.irvine-smith@ninetyone.com}
}
